+++ IL/CHANGELOG.md
@@ -2,6 +2,5 @@
 
 ## [0.1] - 2021-08-31
 
-### Initial version
-- Added minimal set of code to support IL as input
-
+### Added
+- Initial version: minimal set of code to support IL as input