+++ IL/CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - Target setting on IILSession
 
+### Fixed
+- WorkSession.IsIL() check failing if WorkSession is a mock
+
 ## [0.1] - 2021-08-31
 
 ### Added